requirements.txt
python-dotenv==0.19.1
requests==2.25.1
oauth2client==4.1.3
google-api-python-client==2.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
fastapi==0.75.1
uvicorn==0.17.6